requirements.txt
streamlit==1.28.1
Pillow==10.0.0
google-generativeai==0.3.0
PyPDF2==3.0.1
python-docx==0.8.11
tomli==2.0.1
